Moved Dark Mode logic to SidebarV2.jsx, Fixed login screenn (No more sidebar on login, and added sign up conditional to login page)
