+++ app/src/components/SidebarV2.jsx
@@ -1,6 +1,51 @@
+import { useEffect } from 'react'
 import { Link } from 'react-router-dom'
 
 const SidebarV2 = () => {
+  useEffect(() => {
+    const themeToggleDarkIcon = document.getElementById('theme-toggle-dark-icon')
+    const themeToggleLightIcon = document.getElementById('theme-toggle-light-icon')
+
+    if (
+      localStorage.getItem('color-theme') === 'dark' ||
+      (!('color-theme' in localStorage) && window.matchMedia('(prefers-color-scheme: dark)').matches)
+    ) {
+      themeToggleLightIcon.classList.remove('hidden')
+    } else {
+      themeToggleDarkIcon.classList.remove('hidden')
+    }
+
+    const themeToggleBtn = document.getElementById('theme-toggle')
+
+    const handleThemeToggle = () => {
+      themeToggleDarkIcon.classList.toggle('hidden')
+      themeToggleLightIcon.classList.toggle('hidden')
+
+      if (localStorage.getItem('color-theme')) {
+        if (localStorage.getItem('color-theme') === 'light') {
+          document.documentElement.classList.add('dark')
+          localStorage.setItem('color-theme', 'dark')
+        } else {
+          document.documentElement.classList.remove('dark')
+          localStorage.setItem('color-theme', 'light')
+        }
+      } else {
+        if (document.documentElement.classList.contains('dark')) {
+          document.documentElement.classList.remove('dark')
+          localStorage.setItem('color-theme', 'light')
+        } else {
+          document.documentElement.classList.add('dark')
+          localStorage.setItem('color-theme', 'dark')
+        }
+      }
+    }
+
+    themeToggleBtn.addEventListener('click', handleThemeToggle)
+
+    return () => {
+      themeToggleBtn.removeEventListener('click', handleThemeToggle)
+    }
+  }, [])
   return (
     <>
       <nav className='fixed top-0 z-50 w-full bg-white border-b border-slate-200 dark:bg-slate-800 dark:border-slate-700'>
